Rename updater.py to mobject_update_utils.py
