+++ big_ol_pile_of_manim_imports.py
@@ -58,7 +58,7 @@ from manimlib.mobject.three_dimensions import *
 from manimlib.mobject.types.image_mobject import *
 from manimlib.mobject.types.point_cloud_mobject import *
 from manimlib.mobject.types.vectorized_mobject import *
-from manimlib.mobject.updater import *
+from manimlib.mobject.mobject_update_utils import *
 from manimlib.mobject.value_tracker import *
 
 from manimlib.for_3b1b_videos.common_scenes import *
